requirements.txt
requests>=2.31.0
streamlit>=1.34.0
pandas>=2.2.0
google-auth>=2.35.0
google-api-python-client>=2.149.0
